ap-15: Refactor all path elements to own components
