--- app/components/path/PathLine.tsx
@@ -1,7 +1,0 @@
-import type { PathElementProps } from './props';
-
-interface PathLineProps extends PathElementProps {}
-
-export const PathLine: React.FC<PathLineProps> = ({ style }) => {
-  return <View style={style} />;
-};
